(PC-25512)[PRO] refactor: use same custom request for ApiClient
